implemento de Dashboard Vista Principal

+++ src/main/java/com/Andrey/Repositorios/ReservaRepository.java
@@ -1,12 +1,11 @@
 package com.Andrey.Repositorios;
 
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.Andrey.Modelos.Reserva;
 
 @Repository
-public interface ReservaRepository extends JpaRepository<Reserva, Integer> {
+public interface ReservaRepository extends JpaRepository<Reserva, Long> {
 
 }
